+++ README.md
@@ -1,11 +1,11 @@
-# Shisen
+# Shisen OpenCV
 
-[![latest version](https://img.shields.io/github/v/release/ichiro-its/shisen)](https://github.com/ichiro-its/shisen/releases/)
-[![milestone](https://img.shields.io/github/milestones/progress/ichiro-its/shisen/1?label=milestone)](https://github.com/ichiro-its/shisen/milestone/1)
-[![license](https://img.shields.io/github/license/ichiro-its/shisen)](./LICENSE)
-[![test status](https://img.shields.io/github/workflow/status/ichiro-its/shisen/Build%20and%20Test?label=test)](https://github.com/ichiro-its/shisen/actions)
+[![latest version](https://img.shields.io/github/v/release/ichiro-its/shisen_opencv)](https://github.com/ichiro-its/shisen_opencv/releases/)
+[![milestone](https://img.shields.io/github/milestones/progress/ichiro-its/shisen_opencv/1?label=milestone)](https://github.com/ichiro-its/shisen_opencv/milestone/1)
+[![license](https://img.shields.io/github/license/ichiro-its/shisen_opencv)](./LICENSE)
+[![test status](https://img.shields.io/github/workflow/status/ichiro-its/shisen_opencv/Build%20and%20Test?label=test)](https://github.com/ichiro-its/shisen_opencv/actions)
 
-This package provides a camera tools library for [ROS 2](https://docs.ros.org/en/foxy/index.html) projects.
+This package provides an [OpenCV](https://opencv.org/) implementation of [Shisen](https://github.com/ichiro-its/shisen) camera tools project.
 
 ## License
