+++ include/shisen_opencv/camera.hpp
@@ -18,32 +18,58 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#ifndef SHISEN__VIEWER_HPP_
-#define SHISEN__VIEWER_HPP_
+#ifndef SHISEN_OPENCV__CAMERA_HPP_
+#define SHISEN_OPENCV__CAMERA_HPP_
 
+#include <opencv2/videoio.hpp>
 #include <rclcpp/rclcpp.hpp>
 #include <shisen_interfaces/msg/compressed_image.hpp>
+#include <shisen_interfaces/msg/property_event.hpp>
 #include <shisen_interfaces/msg/raw_image.hpp>
+#include <shisen_interfaces/srv/get_properties.hpp>
+#include <shisen_interfaces/srv/set_properties.hpp>
 
+#include <map>
 #include <memory>
 #include <string>
 
-namespace shisen
+namespace shisen_opencv
 {
 
-class Viewer : public rclcpp::Node
+class Camera : public rclcpp::Node
 {
 public:
-  Viewer(std::string node_name, std::string topic_name);
+  explicit Camera(std::string node_name);
+  ~Camera();
+
+  bool open(std::string file_name);
+  bool close();
 
 private:
-  std::shared_ptr<rclcpp::Subscription<shisen_interfaces::msg::RawImage>>
-  raw_image_subscription;
+  static const std::map<std::string, int> property_ids;
 
-  std::shared_ptr<rclcpp::Subscription<shisen_interfaces::msg::CompressedImage>>
-  compressed_image_subscription;
+  cv::VideoCapture video_capture;
+
+  std::map<std::string, double> property_map;
+
+  std::shared_ptr<rclcpp::Publisher<shisen_interfaces::msg::RawImage>>
+  raw_image_publisher;
+
+  std::shared_ptr<rclcpp::Publisher<shisen_interfaces::msg::CompressedImage>>
+  compressed_image_publisher;
+
+  std::shared_ptr<rclcpp::Publisher<shisen_interfaces::msg::PropertyEvent>>
+  property_event_publisher;
+
+  std::shared_ptr<rclcpp::Service<shisen_interfaces::srv::GetProperties>>
+  get_properties_service;
+
+  std::shared_ptr<rclcpp::Service<shisen_interfaces::srv::SetProperties>>
+  set_properties_service;
+
+  std::shared_ptr<rclcpp::TimerBase> capture_timer;
 };
 
-}  // namespace shisen
+}  // namespace shisen_opencv
 
-#endif  // SHISEN__VIEWER_HPP_
+#endif  // SHISEN_OPENCV__CAMERA_HPP_
