+++ include/shisen_opencv/viewer.hpp
@@ -18,58 +18,32 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#ifndef SHISEN__CAMERA_HPP_
-#define SHISEN__CAMERA_HPP_
+#ifndef SHISEN_OPENCV__VIEWER_HPP_
+#define SHISEN_OPENCV__VIEWER_HPP_
 
-#include <opencv2/videoio.hpp>
 #include <rclcpp/rclcpp.hpp>
 #include <shisen_interfaces/msg/compressed_image.hpp>
-#include <shisen_interfaces/msg/property_event.hpp>
 #include <shisen_interfaces/msg/raw_image.hpp>
-#include <shisen_interfaces/srv/get_properties.hpp>
-#include <shisen_interfaces/srv/set_properties.hpp>
 
-#include <map>
 #include <memory>
 #include <string>
 
-namespace shisen
+namespace shisen_opencv
 {
 
-class Camera : public rclcpp::Node
+class Viewer : public rclcpp::Node
 {
 public:
-  explicit Camera(std::string node_name);
-  ~Camera();
-
-  bool open(std::string file_name);
-  bool close();
+  Viewer(std::string node_name, std::string topic_name);
 
 private:
-  static const std::map<std::string, int> property_ids;
+  std::shared_ptr<rclcpp::Subscription<shisen_interfaces::msg::RawImage>>
+  raw_image_subscription;
 
-  cv::VideoCapture video_capture;
-
-  std::map<std::string, double> property_map;
-
-  std::shared_ptr<rclcpp::Publisher<shisen_interfaces::msg::RawImage>>
-  raw_image_publisher;
-
-  std::shared_ptr<rclcpp::Publisher<shisen_interfaces::msg::CompressedImage>>
-  compressed_image_publisher;
-
-  std::shared_ptr<rclcpp::Publisher<shisen_interfaces::msg::PropertyEvent>>
-  property_event_publisher;
-
-  std::shared_ptr<rclcpp::Service<shisen_interfaces::srv::GetProperties>>
-  get_properties_service;
-
-  std::shared_ptr<rclcpp::Service<shisen_interfaces::srv::SetProperties>>
-  set_properties_service;
-
-  std::shared_ptr<rclcpp::TimerBase> capture_timer;
+  std::shared_ptr<rclcpp::Subscription<shisen_interfaces::msg::CompressedImage>>
+  compressed_image_subscription;
 };
 
-}  // namespace shisen
+}  // namespace shisen_opencv
 
-#endif  // SHISEN__CAMERA_HPP_
+#endif  // SHISEN_OPENCV__VIEWER_HPP_
