+++ package.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>shisen</name>
-  <version>0.1.0</version>
-  <description>Camera tools for ROS 2</description>
+  <name>shisen_opencv</name>
+  <version>0.2.0</version>
+  <description>OpenCV implementation of Shisen camera tools project</description>
   <maintainer email="alfi.maulana.f@gmail.com">Alfi Maulana</maintainer>
   <license>MIT License</license>
   <buildtool_depend>ament_cmake_ros</buildtool_depend>
